Rename item.utils → item.structure; use a tmp_path in tests
