+++ item/cli.py
@@ -6,7 +6,7 @@ import click
 from item.model.cli import model
 from item.historical.cli import historical
 from item.remote.cli import remote
-from item.utils import make_template
+from item.structure import make_template
 
 
 @click.group(help=__doc__)
